requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
mysqlclient
gunicorn
flask-cors
python-dotenv